lib: Introduce new fields l3cat_id and mba_id in coreinfo
Right now, both L3 CAT and MBA features use socket as the main identifier.
But this identifier can be different depending on the vendor. Also in
future, there is a possibility the we could have separate ids for each of
these features.

Add two new fields l3cat_id and mba_id in coreinfo for CAT and MBA features
respectively. These fields will be initialized during the initialization. It
is updated based on the vendor.

Right now, l3cat_id and mba_id's are same for both the resources. mba_id
is for future expansion.

Signed-off-by: Babu Moger <babu.moger@amd.com>